+++ scripts/auto-patch.py
@@ -12,6 +12,7 @@ import re
 import smtplib
 import socket
 import subprocess
+from subprocess import CalledProcessError
 import sys
 import tempfile
 from time import sleep
@@ -81,9 +82,128 @@ def logging_add_report(cfg, stream):
 log = logging.getLogger(__name__)
 
 
-class ZypperLockedError(Exception):
-    def __init__(self):
-        super().__init__("ZYPP library is locked")
+class ZypperExitException(CalledProcessError):
+    """Represent a particular non-zero exit code from zypper.
+    """
+    ExitCode = 0
+    Message = None
+    _SubClasses = dict()
+
+    @classmethod
+    def check_returncode(cls, proc):
+        """Raise the appropriate exception if the exit code is non-zero."""
+        if proc.returncode:
+            try:
+                ExcClass = cls._SubClasses[proc.returncode]
+            except KeyError:
+                raise CalledProcessError(proc.returncode, proc.args,
+                                         proc.stdout, proc.stderr) from None
+            raise ExcClass(proc.args, proc.stdout, proc.stderr)
+
+    @classmethod
+    def register_exit_code(cls, subcls):
+        """A class decorator to register the exit code for a subclass.
+        """
+        assert issubclass(subcls, cls)
+        assert subcls.ExitCode and subcls.ExitCode not in cls._SubClasses
+        cls._SubClasses[subcls.ExitCode] = subcls
+        return subcls
+
+    def __init__(self, cmd, stdout=None, stderr=None):
+        if not self.ExitCode:
+            # This is an abstract class that may not be instantiated.
+            # Derived classes must override the class variable
+            # ExitCode.
+            raise NotImplementedError
+        super().__init__(self.ExitCode, cmd, stdout, stderr)
+
+    def __str__(self):
+        if not self.Message:
+            # This is an abstract class.  Derived classes must
+            # override the class variable Message.
+            raise NotImplementedError
+        return self.Message
+
+@ZypperExitException.register_exit_code
+class ZypperBugError(ZypperExitException):
+    ExitCode = 1
+    Message = "Unexpected situation, probably a bug in zypper"
+
+@ZypperExitException.register_exit_code
+class ZypperSyntaxError(ZypperExitException):
+    ExitCode = 2
+    Message = "Syntax error in the zypper call"
+
+@ZypperExitException.register_exit_code
+class ZypperInvalidArgsError(ZypperExitException):
+    ExitCode = 3
+    Message = "Invalid arguments in the zypper call"
+
+@ZypperExitException.register_exit_code
+class ZypperLibraryError(ZypperExitException):
+    ExitCode = 4
+    Message = "Problem reported by ZYPP library"
+
+@ZypperExitException.register_exit_code
+class ZypperPrivilegesError(ZypperExitException):
+    ExitCode = 5
+    Message = "Insufficient privileges calling zypper"
+
+@ZypperExitException.register_exit_code
+class ZypperNoReposError(ZypperExitException):
+    ExitCode = 6
+    Message = "No repositories defined in zypper"
+
+@ZypperExitException.register_exit_code
+class ZypperLockedError(ZypperExitException):
+    ExitCode = 7
+    Message = "ZYPP library is locked"
+
+@ZypperExitException.register_exit_code
+class ZypperCommitError(ZypperExitException):
+    ExitCode = 8
+    Message = "Error during installation or removal of packages"
+
+@ZypperExitException.register_exit_code
+class ZypperPatchesAvailable(ZypperExitException):
+    ExitCode = 100
+    Message = "Patches available for installation"
+
+@ZypperExitException.register_exit_code
+class ZypperSecurityPatchesAvailable(ZypperExitException):
+    ExitCode = 101
+    Message = "Security patches available for installation"
+
+@ZypperExitException.register_exit_code
+class ZypperRebootNeeded(ZypperExitException):
+    ExitCode = 102
+    Message = "Installation of a patch requires reboot"
+
+@ZypperExitException.register_exit_code
+class ZypperRestartNeeded(ZypperExitException):
+    ExitCode = 103
+    Message = "Installation of a patch requires restart of package manager"
+
+@ZypperExitException.register_exit_code
+class ZypperCapabilityNotFound(ZypperExitException):
+    ExitCode = 104
+    Message = ("Arguments does not match available or installed "
+               "package names or capabilities")
+
+@ZypperExitException.register_exit_code
+class ZypperSignal(ZypperExitException):
+    ExitCode = 105
+    Message = "Exit of zypper after receiving a SIGINT or SIGTERM"
+
+@ZypperExitException.register_exit_code
+class ZypperReposSkipped(ZypperExitException):
+    ExitCode = 106
+    Message = "Some repo temporarily disabled because of failure to refresh"
+
+@ZypperExitException.register_exit_code
+class ZypperRPMScriptfailed(ZypperExitException):
+    ExitCode = 107
+    Message = "Some packages install script returned an error"
 
 
 class Zypper:
@@ -91,28 +211,19 @@ class Zypper:
     _zypper = "/usr/bin/zypper"
 
     @classmethod
-    def call(cls, args, stdout=None, retcodes=None):
+    def call(cls, args, stdout=None):
         cmd = [cls._zypper] + args
         log.debug("run: %s", " ".join(cmd))
         stdout.flush()
         proc = subprocess.run(cmd, stdout=stdout, stderr=subprocess.PIPE,
                               universal_newlines=True)
         log.debug("return code from zypper: %d", proc.returncode)
-        if proc.returncode == 7:
-            raise ZypperLockedError()
-        elif (proc.returncode != 0 and
-            not (retcodes and proc.returncode in retcodes)):
-            proc.check_returncode()
-        return proc.returncode
+        ZypperExitException.check_returncode(proc)
 
     @classmethod
     def patch_check(cls, stdout=None):
-        # patch-check
-        # 0: no patches needed
-        # 100: patches available for installation
-        # 101: security patches available for installation
         args = ["--quiet", "--non-interactive", "patch-check"]
-        return cls.call(args, stdout=stdout, retcodes={100, 101})
+        return cls.call(args, stdout=stdout)
 
     @classmethod
     def list_patches(cls, stdout=None):
@@ -121,20 +232,13 @@ class Zypper:
 
     @classmethod
     def patch(cls, stdout=None):
-        # patch: install patches
-        # 0: ok
-        # 102: successful installation, patch requires reboot
-        # 103: successful installation, restart of package manager needed
         args = ["--quiet", "--non-interactive", "patch", "--skip-interactive"]
-        return cls.call(args, stdout=stdout, retcodes={102, 103})
+        return cls.call(args, stdout=stdout)
 
     @classmethod
     def ps(cls, stdout=None):
-        # ps: list processes using deleted files
-        # 0: ok
-        # 102: reboot required
         args = ["--quiet", "ps"]
-        return cls.call(args, stdout=stdout, retcodes={102})
+        return cls.call(args, stdout=stdout)
 
 
 def patch(stdout=None):
@@ -147,9 +251,12 @@ def patch(stdout=None):
         try:
             while True:
                 p = stdout.tell()
-                if Zypper.patch_check(stdout=stdout) == 0:
+                try:
+                    Zypper.patch_check(stdout=stdout)
                     log.debug("no patches needed")
                     break
+                except (ZypperPatchesAvailable, ZypperSecurityPatchesAvailable):
+                    pass
                 stdout.seek(p)
                 m = check_line_pattern.search(stdout.read())
                 if m:
@@ -158,37 +265,35 @@ def patch(stdout=None):
                     log.info("patches are needed")
                 have_patches = True
                 Zypper.list_patches(stdout=stdout)
-                rc = Zypper.patch(stdout=stdout)
-                log.info("patches successfully installed")
-                if rc == 0 or rc == 102:
+                try:
+                    Zypper.patch(stdout=stdout)
+                    log.info("patches successfully installed")
                     break
-                elif rc == 103:
+                except ZypperRebootNeeded:
+                    log.info("patches successfully installed")
+                    break
+                except ZypperRestartNeeded:
                     log.info("patch requires restart to "
                              "check again for more patches")
                     continue
             if not have_patches:
                 return False
-            rc = Zypper.ps(stdout=stdout)
-            if rc == 102:
+            try:
+                Zypper.ps(stdout=stdout)
+            except ZypperRebootNeeded:
                 log.warning("reboot is required after installing patches")
             return True
-        except ZypperLockedError:
+        except (ZypperLockedError, ZypperReposSkipped) as err:
             if try_count < config['retry'].getint('max'):
-                log.info("ZYPP library is locked.  Will try again ...")
+                log.warning("%s.  Will try again ...", err)
                 sleep(config['retry'].getint('wait'))
                 continue
             else:
-                log.error("ZYPP library is locked.  "
-                          "Giving up after %d tries." % try_count)
-                return have_patches
+                err.Message += (".  Giving up after %d tries." % try_count)
+                raise err
 
-def exchandler(type, value, traceback):
-    log.critical("%s: %s", type.__name__, value,
-                 exc_info=(type, value, traceback))
-
-if __name__ == "__main__":
+def main():
     setup_logging(config['logging'])
-    sys.excepthook = exchandler
     with tempfile.TemporaryFile(mode='w+t') as tmpf:
         with logging_add_report(config['logging'], tmpf):
             have_patches = patch(stdout=tmpf)
@@ -205,3 +310,20 @@ if __name__ == "__main__":
                 mailhost = config['mailreport'].get('mailhost')
                 with smtplib.SMTP(mailhost) as smtp:
                     smtp.send_message(msg)
+
+if __name__ == "__main__":
+    try:
+        main()
+    except (ZypperPrivilegesError, ZypperNoReposError, ZypperLockedError,
+            ZypperCommitError, ZypperSignal, ZypperReposSkipped,
+            ZypperRPMScriptfailed) as err:
+        log.error(err)
+        sys.exit(err.ExitCode)
+    except ZypperExitException as err:
+        log.critical("Internal error %s: %s", type(err).__name__, err,
+                     exc_info=err)
+        sys.exit(err.ExitCode)
+    except Exception as err:
+        log.critical("Internal error %s: %s", type(err).__name__, err,
+                     exc_info=err)
+        sys.exit(-1)